Extend setup for Ubuntu 20.04 LTS
The Newcomer's Guide is extended with the instructions to set up
the development environment on Ubuntu 20.14 LTS. Additionally a
link is provided to an already set-up VM, which unfortunately
can only be used by Hyper-V. A comment was also added to the
Windows 10 WSL section detailing why it can't be used as a dev.
env. for this project.